requirements.txt
openai>=1.30.0
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0
gTTS>=2.5.0
playsound==1.2.2
python-dotenv>=1.0.0
